Add base class for data objects
